V2.0 集成Redis

+++ dubbo-provider/src/main/java/com/demo/mapper/UserMapper.java
@@ -3,9 +3,6 @@ package com.demo.mapper;
 import com.demo.entity.User;
 
 public interface UserMapper {
-
-
-
-    User selectByUserId(Integer userId);
+    User selectByUserId(Long userId);
 
 }
